+++ src/screens/SearchScreen.js
@@ -1,0 +1,20 @@
+import React from 'react'
+import { View, Text, StyleSheet } from 'react-native'
+
+const SearchScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Search Screen</Text>
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+})
+
+export default SearchScreen
